Address PR #1818 review comments
- Extract query.service.ts's free-standing helpers into api/src/util/queryFanout.ts
  (parentId fan-out + result merging) and querySelector.ts (selector field
  extraction), shrinking the service file and matching the existing util/ layout.
- Trim the duplicated fan-out comments in PinnedTopics/HomePagePinned/PinnedVideo
  that restated queryIntrospection.ts's internals instead of pointing to it.
- Centralize the Post-vs-Tag publishDate-windowing check into a single
  isWindowedContentSubType() predicate in shared/config.ts, replacing four
  independent `=== DocType.Post` comparisons in sync.ts, liveSync.ts, and
  retention.ts so the policy can't drift between call sites. No behavior change.
